# [C#] Bring matmul.cs in line with the matmul implementation of other languages