The current version of `matmul.cs` does not function the same as the other versions because it uses a multidimensional array instead of an array of arrays.

This is also much slower in C#. So this PR brings a speedup of over 2x. This brings the performance much closer to the java version.

On my machine (Mac Mini M1 16Gb):
Before: **5.458 s** ±  0.023 s
After: **2.112 s** ±  0.015 s