@attractivechaos 2024-01-04
Thank you! Figure updated.